# internal/routes/handler.go
package routes

import (
	"context"
	"database/sql"
	"fmt"
	"github.com/dgrijalva/jwt-go"
	"log"
	"net/http"
	"time"
)

type Handler struct {
	DB *sql.DB
}

type User struct {
	// Пользователь
	ID       string  `json:"id"`
	Name     string  `json:"name"`
	Password string  `json:"password"`
	IsFriend *string `json:"is_friend,omitempty"`
}

type Room struct {
	// Комната
	ID        string `json:"id"`
	Name      string `json:"name"`
	CreatedBy string `json:"created_by"`
}

func NewHandler(db *sql.DB) *Handler {
	return &Handler{
		DB: db,
	}
}

var jwtSecret = []byte("myapp-super-secret-jwt-key-2024-development")

func AuthMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		authHeader := r.Header.Get("Authorization")
		if authHeader == "" {
			http.Error(w, "Authorization header is required", http.StatusUnauthorized)
			return
		}

		// Формат заголовка: "Bearer {token}"
		if len(authHeader) < 7 || authHeader[:7] != "Bearer " {
			http.Error(w, "Invalid authorization format", http.StatusUnauthorized)
			return
		}

		tokenString := authHeader[7:]

		// Парсинг и проверка токена
		token, err := jwt.Parse(tokenString, func(token *jwt.Token) (interface{}, error) {
			// Проверяем, что алгоритм подписи соответствует ожидаемому
			if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
				return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
			}
			return jwtSecret, nil
		})

		if err != nil {
			log.Printf("Error parsing token: %v", err)
			http.Error(w, "Invalid token", http.StatusUnauthorized)
			return
		}

		if !token.Valid {
			http.Error(w, "Invalid token", http.StatusUnauthorized)
			return
		}

		// Извлечение ID пользователя из токена
		claims, ok := token.Claims.(jwt.MapClaims)
		if !ok {
			http.Error(w, "Invalid token claims", http.StatusUnauthorized)
			return
		}

		userID, ok := claims["user_id"].(string)
		if !ok {
			log.Printf("AuthMiddleware: Invalid user ID in token, claims: %+v", claims)
			http.Error(w, "Invalid user ID in token", http.StatusUnauthorized)
			return
		}

		// Добавление ID пользователя в контекст запроса
		ctx := context.WithValue(r.Context(), "user_id", string(userID))
		next.ServeHTTP(w, r.WithContext(ctx))
	})
}

func createToken(userID string, userName string) (string, error) {
	claims := jwt.MapClaims{
		"user_id":   userID,
		"user_name": userName,
		"exp":       time.Now().Add(time.Hour * 24).Unix(), // Токен действителен 24 часа
	}

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
	return token.SignedString(jwtSecret)
}